Do matching with NS_XXX errors with "NS_XXX" error names correctly.
On old Firefox, an NS_XXX error object has a human readable message as its ".message".
However, on Firefox 31, the object has no message and its ".message" is a blank string.
As the result, comparison of NS_XXX's ".message" with "Cr.NS_OK" (an integer "0") was unexpectedly detected as a same error.
To avoid such a wrong detection, we have to compare the message and the error code as strings.